Partially fix tests
